add scope to view only owned articles

+++ app/Http/Controllers/ArticleController.php
@@ -46,6 +46,7 @@ class ArticleController extends Controller
 	}
 
 	public function edit(Article $article){
+		$this->authorize('own', $article);
 
 		return Inertia::render('Article/Edit',[
 			'article' => $article
@@ -53,6 +54,8 @@ class ArticleController extends Controller
 	}
 
 	public function update(Article $article, Request $request){
+		$this->authorize('own', $article);
+
 		if($this->service->update($article, $request)){
 			return redirect()->route('article.index')->with('success', 'Create article successfully');
 		}
@@ -60,6 +63,8 @@ class ArticleController extends Controller
 
 	}
 	public function destroy(Article $article){
+		$this->authorize('own', $article);
+
 		$article->delete();
 
 		return redirect()->back()->with('success', 'Delete article successfully');
